requirements.txt
--extra-index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments
yfinance
pandas
pandas-ta
aiohttp
numpy
python-dotenv
quantstats
streamlit
plotly
sentry-sdk
matplotlib
ta
talib-binary
